Refactor Package Names
Namespace had to be refactored to match my newly acquired domain name (nixxcode.com)
