+++ jvmbrotli/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.nixx.jvmbrotli</groupId>
+        <groupId>com.nixxcode.jvmbrotli</groupId>
         <artifactId>jvmbrotli-parent</artifactId>
         <version>0.1-SNAPSHOT</version>
     </parent>
@@ -15,10 +15,27 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.nixx.jvmbrotli</groupId>
+            <groupId>com.nixxcode.jvmbrotli</groupId>
             <artifactId>jvmbrotli-win32-x86-amd64</artifactId>
             <version>0.1-SNAPSHOT</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>2.6</version>
+                <configuration>
+                    <archive>
+                        <manifest>
+                            <addDefaultImplementationEntries>true</addDefaultImplementationEntries>
+                        </manifest>
+                    </archive>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
