+++ natives/src/main/cpp/decoder_jni.cc
@@ -41,7 +41,7 @@ extern "C" {
  * @returns direct ByteBuffer if directBufferSize is not 0; otherwise null
  */
 JNIEXPORT jobject JNICALL
-Java_org_brotli_wrapper_dec_DecoderJNI_nativeCreate(
+Java_com_nixxcode_jvmbrotli_dec_DecoderJNI_nativeCreate(
     JNIEnv* env, jobject /*jobj*/, jlongArray ctx) {
   bool ok = true;
   DecoderHandle* handle = nullptr;
@@ -104,7 +104,7 @@ Java_org_brotli_wrapper_dec_DecoderJNI_nativeCreate(
  *                     0 to process further previous input
  */
 JNIEXPORT void JNICALL
-Java_org_brotli_wrapper_dec_DecoderJNI_nativePush(
+Java_com_nixxcode_jvmbrotli_dec_DecoderJNI_nativePush(
     JNIEnv* env, jobject /*jobj*/, jlongArray ctx, jint input_length) {
   jlong context[3];
   env->GetLongArrayRegion(ctx, 0, 3, context);
@@ -159,7 +159,7 @@ Java_org_brotli_wrapper_dec_DecoderJNI_nativePush(
  *          any further invocation; null in case of error
  */
 JNIEXPORT jobject JNICALL
-Java_org_brotli_wrapper_dec_DecoderJNI_nativePull(
+Java_com_nixxcode_jvmbrotli_dec_DecoderJNI_nativePull(
     JNIEnv* env, jobject /*jobj*/, jlongArray ctx) {
   jlong context[3];
   env->GetLongArrayRegion(ctx, 0, 3, context);
@@ -187,7 +187,7 @@ Java_org_brotli_wrapper_dec_DecoderJNI_nativePull(
  * @param ctx {in_cookie} tuple
  */
 JNIEXPORT void JNICALL
-Java_org_brotli_wrapper_dec_DecoderJNI_nativeDestroy(
+Java_com_nixxcode_jvmbrotli_dec_DecoderJNI_nativeDestroy(
     JNIEnv* env, jobject /*jobj*/, jlongArray ctx) {
   jlong context[3];
   env->GetLongArrayRegion(ctx, 0, 3, context);
