+++ natives/win32-x86-amd64/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.nixx.jvmbrotli</groupId>
+        <groupId>com.nixxcode.jvmbrotli</groupId>
         <artifactId>jvmbrotli-natives</artifactId>
         <version>0.1-SNAPSHOT</version>
     </parent>
